backend/requirements.txt
Flask
Flask-CORS
pymongo
twilio
python-decouple
python-dotenv
gunicorn
bcrypt
pyjwt
